+++ Runtime/Scripts/Rendering/IPointCloudRenderer.cs
@@ -5,10 +5,11 @@ namespace BuildingVolumes.Player
 {
   public interface IPointCloudRenderer
   {
-    public void Setup(SequenceConfiguration configuration, Transform parent, float pointSize, float emission, Material pointcloudMaterial, bool instantiateMaterial);
+    public void Setup(SequenceConfiguration configuration, Transform parent, PointcloudRenderSettings settings);
     public void SetFrame(Frame frame);
     public void SetPointSize(float size);
     public void SetPointEmission(float emission);
+    public void SetOpacity(float opacity);
     public void SetPointcloudMaterial(Material pointcloudMaterial, bool instantiateMaterial);
     public void SetPointcloudMaterial(Material pointcloudMaterial, float pointSize, float pointEmission, bool instantiateMaterial);
 
@@ -19,4 +20,3 @@ namespace BuildingVolumes.Player
     public bool IsDisposed();
   }
 }
-